added menu v1.0

+++ src/main/java/Speise.java
@@ -1,19 +1,22 @@
 public class Speise {
-    private int preis;
+    private double preis;
     private String speise;
 
-    public Speise(int preis, String speise) {
+    public Speise(double preis, String speise) {
         this.preis = preis;
         this.speise = speise;
     }
 
-    public int getPreis() {
+
+    public double getPreis() {
         return preis;
     }
 
-    public void setPreis(int preis) {
+
+    public void setPreis(double preis) {
         this.preis = preis;
     }
+
 
     public String getSpeise() {
         return speise;
